+++ lib/algorithms/mod.rs
@@ -1,2 +1,3 @@
 pub mod breadth_first_search;
 pub mod dijkstra;
+pub mod path;