+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.3.0 (2024-12-04)
+
+### Feat
+
+- implement path data types
+
+### Fix
+
+- **algorithms**: remove illegal moves from distance map/flow field
+
 ## 0.2.1 (2024-11-25)
 
 ### Fix